FY-19/refactor: separate class definition files
